put changes in util
