+++ coffea/lookup_tools/jec_uncertainty_lookup.py
@@ -8,12 +8,9 @@ import scipy
 import scipy.interpolate
 from scipy.interpolate import interp1d
 
-from .jme_standard_function import searchsorted_wrapped
-
-from ..util import USE_CUPY
+from ..util import USE_CUPY, interp1d, searchsorted_wrapped
 
 if USE_CUPY:
-    import cupy as np
     import cupy
 
 def masked_bin_eval(dim1_indices, dimN_bins, dimN_vals):
@@ -26,28 +23,6 @@ def masked_bin_eval(dim1_indices, dimN_bins, dimN_vals):
                                     0, len(dimN_bins[i]) - 2)
     return dimN_indices
 
-if USE_CUPY:
-    class interp1d:
-        def __init__(self, x, y):
-            self.x = cupy.array(x)
-            self.y = cupy.array(y)
-            self.ydiffs = cupy.diff(self.y)
-            self.xdiffs = cupy.diff(self.x)
-            self.delta = self.ydiffs / self.xdiffs
-    
-            #self.scipy_interp1d = scipy.interpolate.interp1d(x, y)
-    
-        def __call__(self, xvals):
-            idx = searchsorted_wrapped(self.x, xvals, asnumpy=False)
-           
-            xshifts =  self.x[idx] - xvals 
-            yvals = self.y[idx] 
-            yvals_new = yvals - self.delta[idx-1]*xshifts
-            
-            #yvals_scipy = self.scipy_interp1d(xvals)
-            #assert(np.sum(np.power(yvals_new - yvals_scipy,2)) < 1e-1)
-            return yvals_new
- 
 class jec_uncertainty_lookup(lookup_base):
     """
     This class defines a lookup table for jet energy scale uncertainties.
@@ -109,25 +84,23 @@ class jec_uncertainty_lookup(lookup_base):
         bin_vals = {argname: args[self._dim_args[argname]] for argname in self._dim_order}
         eval_vals = {argname: args[self._eval_args[argname]] for argname in self._eval_vars}
         
-        #if USE_CUPY:
-        #    bin_vals = {k: cupy.array(v) for k, v in bin_vals.items()}
-        #    eval_vals = {k: cupy.array(v) for k, v in eval_vals.items()}
+        _np = np
+        if USE_CUPY:
+            _np = cupy
 
         # lookup the bins that we care about
         dim1_name = self._dim_order[0]
-        dim1_indices = np.clip(searchsorted_wrapped(self._bins[dim1_name],
+        dim1_indices = _np.clip(searchsorted_wrapped(self._bins[dim1_name],
                                                bin_vals[dim1_name],
                                                side='right', asnumpy=not USE_CUPY) - 1,
                                0, self._bins[dim1_name].size - 2)
-        #if USE_CUPY:
-        #    dim1_indices = cupy.array(dim1_indices)
 
         # get clamp values and clip the inputs
-        outs = np.ones(shape=(args[0].size, 2), dtype=np.float)
-        for i in np.unique(dim1_indices):
+        outs = _np.ones(shape=(args[0].size, 2), dtype=np.float)
+        for i in _np.unique(dim1_indices):
             i = int(i)
-            mask = np.where(dim1_indices == i)
-            vals = np.clip(eval_vals[self._eval_vars[0]][mask],
+            mask = _np.where(dim1_indices == i)
+            vals = _np.clip(eval_vals[self._eval_vars[0]][mask],
                            self._eval_knots[0], self._eval_knots[-1])
             outs[:, 0][mask] += self._eval_ups[i](vals)
             outs[:, 1][mask] -= self._eval_downs[i](vals)
