+++ coffea/util.py
@@ -33,3 +33,45 @@ def save(output, filename):
 
 import os
 USE_CUPY = int(os.environ.get("HEPACCELERATE_CUDA", 0)) == 1
+numpy_backend = numpy
+
+if USE_CUPY:
+    import cupy
+    numpy_backend = cupy
+    from hepaccelerate.backend_cuda import searchsorted
+else:
+    from hepaccelerate.backend_cpu import searchsorted
+
+def searchsorted_wrapped(arr, vals, side="left", asnumpy=True):
+    if USE_CUPY:
+        arr = cupy.array(arr, dtype=cupy.float32)
+        vals = cupy.array(vals, dtype=cupy.float32)
+
+    idx = searchsorted(arr, vals, side)
+
+    if USE_CUPY and asnumpy:
+        idx = cupy.asnumpy(idx)
+    assert(np.all(idx < len(arr))) 
+    assert(np.all(idx >= 0)) 
+    return idx
+
+class interp1d:
+    def __init__(self, x, y):
+        self.x = numpy_backend.array(x)
+        self.y = numpy_backend.array(y)
+        self.ydiffs = numpy_backend.diff(self.y)
+        self.xdiffs = numpy_backend.diff(self.x)
+        self.delta = self.ydiffs / self.xdiffs
+
+        #self.scipy_interp1d = scipy.interpolate.interp1d(x, y)
+
+    def __call__(self, xvals):
+        idx = searchsorted_wrapped(self.x, xvals, asnumpy=False)
+       
+        xshifts =  self.x[idx] - xvals 
+        yvals = self.y[idx] 
+        yvals_new = yvals - self.delta[idx-1]*xshifts
+        
+        #yvals_scipy = self.scipy_interp1d(xvals)
+        #assert(np.sum(np.power(yvals_new - yvals_scipy,2)) < 1e-1)
+        return yvals_new
